feat: implement comprehensive site auditing system with automated checks and dashboard integration

+++ supabase/migrations/20260401000001_vitals_functions.sql
@@ -8,9 +8,9 @@ create or replace function get_domain_vitals_p75(
   p_until timestamptz
 )
 returns table (
-  lcp numeric,
-  cls numeric,
-  inp numeric,
+  lcp double precision,
+  cls double precision,
+  inp double precision,
   sample_size bigint
 ) language plpgsql security definer as $$
 begin
